refactor(python): move stellar-sdk based code to trezorlib.stellar_sdk_helpers
The new module imports stellar-sdk unconditionally, so pyright can narrow
isinstance() checks, which it refuses for names bound in a try/except
ImportError block. trezorlib.stellar keeps device communication only.
