+++ python/src/trezorlib/stellar_sdk_helpers.py
@@ -1,0 +1,698 @@
+# This file is part of the Trezor project.
+#
+# Copyright (C) SatoshiLabs and contributors
+#
+# This library is free software: you can redistribute it and/or modify
+# it under the terms of the GNU Lesser General Public License version 3
+# as published by the Free Software Foundation.
+#
+# This library is distributed in the hope that it will be useful,
+# but WITHOUT ANY WARRANTY; without even the implied warranty of
+# MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+# GNU Lesser General Public License for more details.
+#
+# You should have received a copy of the License along with this library.
+# If not, see <https://www.gnu.org/licenses/lgpl-3.0.html>.
+
+"""Translation of stellar-sdk objects into Trezor protobuf messages.
+
+Requires the optional `stellar-sdk` dependency (`pip install trezor[stellar]`)
+and fails to import without it. Device communication lives in `trezorlib.stellar`.
+"""
+
+from __future__ import annotations
+
+import typing as t
+from collections.abc import Iterable, Iterator
+
+from stellar_sdk import (
+    AccountMerge,
+    AllowTrust,
+    Asset,
+    BumpSequence,
+    ChangeTrust,
+    ClaimClaimableBalance,
+    CreateAccount,
+    CreatePassiveSellOffer,
+    HashMemo,
+    IdMemo,
+    InvokeHostFunction,
+    LiquidityPoolAsset,
+    ManageBuyOffer,
+    ManageData,
+    ManageSellOffer,
+    MuxedAccount,
+    NoneMemo,
+    Operation,
+    PathPaymentStrictReceive,
+    PathPaymentStrictSend,
+    Payment,
+    ReturnHashMemo,
+    SetOptions,
+    TextMemo,
+    TransactionEnvelope,
+    TrustLineEntryFlag,
+    xdr,
+)
+from stellar_sdk import Address as StellarAddress
+
+from . import messages
+
+if t.TYPE_CHECKING:
+    from .stellar import StellarMessageType
+
+# Protocol 27 XDR (SOROBAN_CREDENTIALS_ADDRESS_V2 etc.) is only available
+# in stellar-sdk >= 15; on Python 3.9 the `stellar` extra installs
+# stellar-sdk 13 (see pyproject.toml), which lacks it.
+HAVE_STELLAR_SDK_PROTOCOL_27 = hasattr(
+    xdr.SorobanCredentialsType, "SOROBAN_CREDENTIALS_ADDRESS_V2"
+)
+
+
+def from_envelope(
+    envelope: TransactionEnvelope,
+    asset_hints: Iterable[Asset] = (),
+) -> tuple[messages.StellarSignTx, list[StellarMessageType], messages.StellarTxExt]:
+    """Parse a transaction envelope into a tuple of:
+
+    tx - a StellarSignTx describing the transaction header
+    operations - a list of protobuf messages, one per operation
+    tx_ext - a StellarTxExt describing the transaction extension: v=1 carrying
+        the Soroban data for Soroban transactions, otherwise v=0
+
+    Each asset in `asset_hints` is attached to the contract invocations that
+    target its Stellar Asset Contract, letting the device present them as token
+    operations. Hints matching no invocation are ignored. The device re-derives
+    the address before using a hint and falls back to the raw contract UI if it
+    does not match.
+    """
+
+    parsed_tx = envelope.transaction
+    if parsed_tx.preconditions is None or parsed_tx.preconditions.time_bounds is None:
+        raise ValueError("Timebounds are mandatory")
+
+    memo_type = messages.StellarMemoType.NONE
+    memo_text = None
+    memo_id = None
+    memo_hash = None
+    if isinstance(parsed_tx.memo, NoneMemo):
+        pass
+    elif isinstance(parsed_tx.memo, TextMemo):
+        # memo_text is specified as UTF-8 string, but returned as bytes from the XDR parser
+        memo_type = messages.StellarMemoType.TEXT
+        memo_text = parsed_tx.memo.memo_text.decode("utf-8")
+    elif isinstance(parsed_tx.memo, IdMemo):
+        memo_type = messages.StellarMemoType.ID
+        memo_id = parsed_tx.memo.memo_id
+    elif isinstance(parsed_tx.memo, HashMemo):
+        memo_type = messages.StellarMemoType.HASH
+        memo_hash = parsed_tx.memo.memo_hash
+    elif isinstance(parsed_tx.memo, ReturnHashMemo):
+        memo_type = messages.StellarMemoType.RETURN
+        memo_hash = parsed_tx.memo.memo_return
+    else:
+        raise ValueError("Unsupported memo type")
+
+    _raise_if_account_muxed_id_exists(parsed_tx.source)
+    tx = messages.StellarSignTx(
+        source_account=parsed_tx.source.account_id,
+        fee=parsed_tx.fee,
+        sequence_number=parsed_tx.sequence,
+        timebounds_start=parsed_tx.preconditions.time_bounds.min_time,
+        timebounds_end=parsed_tx.preconditions.time_bounds.max_time,
+        memo_type=memo_type,
+        memo_text=memo_text,
+        memo_id=memo_id,
+        memo_hash=memo_hash,
+        num_operations=len(parsed_tx.operations),
+        network_passphrase=envelope.network_passphrase,
+    )
+
+    operations = [_read_operation(op) for op in parsed_tx.operations]
+
+    asset_hints = list(asset_hints)
+    if asset_hints:
+        sacs = _sac_addresses(asset_hints, envelope.network_passphrase)
+        for op in operations:
+            for args in _operation_contract_args(op):
+                args.asset_hint = sacs.get(args.contract_address)
+
+    if parsed_tx.soroban_data:
+        tx_ext = messages.StellarTxExt(
+            v=1,
+            soroban_data=parsed_tx.soroban_data.to_xdr_bytes(),
+        )
+    else:
+        tx_ext = messages.StellarTxExt(v=0)
+
+    return tx, operations, tx_ext
+
+
+def from_authorization_entry(
+    entry: xdr.SorobanAuthorizationEntry,
+    network_passphrase: str | None = None,
+    asset_hints: Iterable[Asset] = (),
+) -> messages.StellarSorobanAuthorizationWithAddress:
+    """Translate a Soroban authorization entry into its signing request payload.
+
+    The resulting message carries exactly the fields committed into the
+    entry's authorization payload (the WITH_ADDRESS preimage of Protocol 27).
+    Only SOROBAN_CREDENTIALS_ADDRESS_V2 and
+    SOROBAN_CREDENTIALS_ADDRESS_WITH_DELEGATES entries are supported.
+
+    For entries with delegates the payload is bound to the entry's top-level
+    address: under CAP-71-01 every delegated signer, however deeply nested,
+    authenticates that same payload and context, so one request serves the
+    top-level account's own signer and any of its delegates -- `address_n`
+    decides which key signs.
+
+    See `from_envelope` for `asset_hints`. Matching them needs the network the
+    entry is signed for, so `network_passphrase` is required alongside them.
+    """
+    credentials = None
+    if HAVE_STELLAR_SDK_PROTOCOL_27:
+        if (
+            entry.credentials.type
+            == xdr.SorobanCredentialsType.SOROBAN_CREDENTIALS_ADDRESS_V2
+        ):
+            assert entry.credentials.address_v2 is not None
+            credentials = entry.credentials.address_v2
+        elif (
+            entry.credentials.type
+            == xdr.SorobanCredentialsType.SOROBAN_CREDENTIALS_ADDRESS_WITH_DELEGATES
+        ):
+            assert entry.credentials.address_with_delegates is not None
+            credentials = entry.credentials.address_with_delegates.address_credentials
+    if credentials is None:
+        raise ValueError(
+            f"Unsupported SorobanCredentials type: {entry.credentials.type}"
+        )
+
+    invocation = _read_authorized_invocation(entry.root_invocation)
+    asset_hints = list(asset_hints)
+    if asset_hints:
+        if network_passphrase is None:
+            raise ValueError("network_passphrase is required to match asset hints")
+        sacs = _sac_addresses(asset_hints, network_passphrase)
+        for args in _invocation_contract_args(invocation):
+            args.asset_hint = sacs.get(args.contract_address)
+
+    return messages.StellarSorobanAuthorizationWithAddress(
+        nonce=credentials.nonce.int64,
+        signature_expiration_ledger=credentials.signature_expiration_ledger.uint32,
+        address=_read_sc_address(credentials.address),
+        invocation=invocation,
+    )
+
+
+def parse_asset(value: str) -> Asset:
+    """Parse an asset written the SEP-11 way: `native`, or CODE:ISSUER."""
+    if value == "native":
+        return Asset.native()
+    code, _, issuer = value.partition(":")
+    if not issuer:
+        raise ValueError("expected CODE:ISSUER or `native`")
+    return Asset(code, issuer)
+
+
+def _sac_addresses(
+    asset_hints: Iterable[Asset], network_passphrase: str
+) -> dict[str, messages.StellarAsset]:
+    """Index the hinted assets by the address of their Stellar Asset Contract."""
+    return {
+        asset.contract_id(network_passphrase): _read_asset(asset)
+        for asset in asset_hints
+    }
+
+
+def _invocation_contract_args(
+    invocation: messages.StellarSorobanAuthorizedInvocation,
+) -> Iterator[messages.StellarInvokeContractArgs]:
+    """Walk the contract invocations of an authorized invocation tree."""
+    if invocation.function.contract_fn is not None:
+        yield invocation.function.contract_fn
+    for sub in invocation.sub_invocations:
+        yield from _invocation_contract_args(sub)
+
+
+def _operation_contract_args(
+    op: StellarMessageType,
+) -> Iterator[messages.StellarInvokeContractArgs]:
+    """Walk the contract invocations an operation displays: the host function
+    it invokes and the authorization trees it carries."""
+    if not isinstance(op, messages.StellarInvokeHostFunctionOp):
+        return
+    if op.function.invoke_contract is not None:
+        yield op.function.invoke_contract
+    for entry in op.auth:
+        yield from _invocation_contract_args(entry.root_invocation)
+
+
+def _read_operation(op: Operation) -> StellarMessageType:
+    # TODO: Let's add muxed account support later.
+    if op.source:
+        _raise_if_account_muxed_id_exists(op.source)
+        source_account = op.source.account_id
+    else:
+        source_account = None
+    if isinstance(op, CreateAccount):
+        return messages.StellarCreateAccountOp(
+            source_account=source_account,
+            new_account=op.destination,
+            starting_balance=_read_amount(op.starting_balance),
+        )
+    if isinstance(op, Payment):
+        _raise_if_account_muxed_id_exists(op.destination)
+        return messages.StellarPaymentOp(
+            source_account=source_account,
+            destination_account=op.destination.account_id,
+            asset=_read_asset(op.asset),
+            amount=_read_amount(op.amount),
+        )
+    if isinstance(op, PathPaymentStrictReceive):
+        _raise_if_account_muxed_id_exists(op.destination)
+        return messages.StellarPathPaymentStrictReceiveOp(
+            source_account=source_account,
+            send_asset=_read_asset(op.send_asset),
+            send_max=_read_amount(op.send_max),
+            destination_account=op.destination.account_id,
+            destination_asset=_read_asset(op.dest_asset),
+            destination_amount=_read_amount(op.dest_amount),
+            paths=[_read_asset(asset) for asset in op.path],
+        )
+    if isinstance(op, ManageSellOffer):
+        return messages.StellarManageSellOfferOp(
+            source_account=source_account,
+            selling_asset=_read_asset(op.selling),
+            buying_asset=_read_asset(op.buying),
+            amount=_read_amount(op.amount),
+            price_n=op.price.n,
+            price_d=op.price.d,
+            offer_id=op.offer_id,
+        )
+    if isinstance(op, CreatePassiveSellOffer):
+        return messages.StellarCreatePassiveSellOfferOp(
+            source_account=source_account,
+            selling_asset=_read_asset(op.selling),
+            buying_asset=_read_asset(op.buying),
+            amount=_read_amount(op.amount),
+            price_n=op.price.n,
+            price_d=op.price.d,
+        )
+    if isinstance(op, SetOptions):
+        operation = messages.StellarSetOptionsOp(
+            source_account=source_account,
+            inflation_destination_account=op.inflation_dest,
+            clear_flags=op.clear_flags,
+            set_flags=op.set_flags,
+            master_weight=op.master_weight,
+            low_threshold=op.low_threshold,
+            medium_threshold=op.med_threshold,
+            high_threshold=op.high_threshold,
+            home_domain=op.home_domain,
+        )
+        if op.signer:
+            signer_type = op.signer.signer_key.signer_key_type
+            operation.signer_type = messages.StellarSignerType(signer_type.value)
+            operation.signer_key = op.signer.signer_key.signer_key
+            operation.signer_weight = op.signer.weight
+        return operation
+    if isinstance(op, ChangeTrust):
+        if isinstance(op.asset, LiquidityPoolAsset):
+            raise ValueError("Liquidity pool assets are not supported")
+        return messages.StellarChangeTrustOp(
+            source_account=source_account,
+            asset=_read_asset(op.asset),
+            limit=_read_amount(op.limit),
+        )
+    if isinstance(op, AllowTrust):
+        if op.authorize not in (
+            TrustLineEntryFlag.UNAUTHORIZED_FLAG,
+            TrustLineEntryFlag.AUTHORIZED_FLAG,
+        ):
+            raise ValueError("Unsupported trust line flag")
+        asset_type = (
+            messages.StellarAssetType.ALPHANUM4
+            if len(op.asset_code) <= 4
+            else messages.StellarAssetType.ALPHANUM12
+        )
+        return messages.StellarAllowTrustOp(
+            source_account=source_account,
+            trusted_account=op.trustor,
+            asset_type=asset_type,
+            asset_code=op.asset_code,
+            is_authorized=bool(op.authorize.value),
+        )
+    if isinstance(op, AccountMerge):
+        _raise_if_account_muxed_id_exists(op.destination)
+        return messages.StellarAccountMergeOp(
+            source_account=source_account,
+            destination_account=op.destination.account_id,
+        )
+    # Inflation is not implemented since anyone can submit this operation to the network
+    if isinstance(op, ManageData):
+        return messages.StellarManageDataOp(
+            source_account=source_account,
+            key=op.data_name,
+            value=op.data_value,
+        )
+    if isinstance(op, BumpSequence):
+        return messages.StellarBumpSequenceOp(
+            source_account=source_account, bump_to=op.bump_to
+        )
+    if isinstance(op, ManageBuyOffer):
+        return messages.StellarManageBuyOfferOp(
+            source_account=source_account,
+            selling_asset=_read_asset(op.selling),
+            buying_asset=_read_asset(op.buying),
+            amount=_read_amount(op.amount),
+            price_n=op.price.n,
+            price_d=op.price.d,
+            offer_id=op.offer_id,
+        )
+    if isinstance(op, PathPaymentStrictSend):
+        _raise_if_account_muxed_id_exists(op.destination)
+        return messages.StellarPathPaymentStrictSendOp(
+            source_account=source_account,
+            send_asset=_read_asset(op.send_asset),
+            send_amount=_read_amount(op.send_amount),
+            destination_account=op.destination.account_id,
+            destination_asset=_read_asset(op.dest_asset),
+            destination_min=_read_amount(op.dest_min),
+            paths=[_read_asset(asset) for asset in op.path],
+        )
+    if isinstance(op, ClaimClaimableBalance):
+        return messages.StellarClaimClaimableBalanceOp(
+            source_account=source_account,
+            balance_id=bytes.fromhex(op.balance_id),
+        )
+    if isinstance(op, InvokeHostFunction):
+        return messages.StellarInvokeHostFunctionOp(
+            source_account=source_account,
+            function=_read_host_function(op.host_function),
+            auth=[_read_authorization_entry(entry) for entry in op.auth],
+        )
+    raise ValueError(f"Unknown operation type: {op.__class__.__name__}")
+
+
+def _raise_if_account_muxed_id_exists(account: MuxedAccount) -> None:
+    # Currently Trezor firmware does not support MuxedAccount,
+    # so we throw an exception here.
+    if account.account_muxed_id is not None:
+        raise ValueError("MuxedAccount is not supported")
+
+
+def _read_amount(amount: str) -> int:
+    return Operation.to_xdr_amount(amount)
+
+
+def _read_asset(asset: Asset) -> messages.StellarAsset:
+    """Reads a stellar Asset from unpacker"""
+    if asset.is_native():
+        return messages.StellarAsset(type=messages.StellarAssetType.NATIVE)
+    if asset.guess_asset_type() == "credit_alphanum4":
+        return messages.StellarAsset(
+            type=messages.StellarAssetType.ALPHANUM4,
+            code=asset.code,
+            issuer=asset.issuer,
+        )
+    if asset.guess_asset_type() == "credit_alphanum12":
+        return messages.StellarAsset(
+            type=messages.StellarAssetType.ALPHANUM12,
+            code=asset.code,
+            issuer=asset.issuer,
+        )
+    raise ValueError("Unsupported asset type")
+
+
+def _read_sc_address(address: xdr.SCAddress) -> str:
+    """Read an SCAddress from XDR."""
+    addr = StellarAddress.from_xdr_sc_address(address)
+    return addr.address
+
+
+def _read_sc_val(val: xdr.SCVal) -> messages.StellarSCVal:
+    """Read an SCVal from XDR."""
+    if val.type == xdr.SCValType.SCV_BOOL:
+        return messages.StellarSCVal(type=messages.StellarSCValType.SCV_BOOL, b=val.b)
+    elif val.type == xdr.SCValType.SCV_VOID:
+        return messages.StellarSCVal(type=messages.StellarSCValType.SCV_VOID)
+    elif val.type == xdr.SCValType.SCV_U32:
+        assert val.u32 is not None
+        return messages.StellarSCVal(
+            type=messages.StellarSCValType.SCV_U32, u32=val.u32.uint32
+        )
+    elif val.type == xdr.SCValType.SCV_I32:
+        assert val.i32 is not None
+        return messages.StellarSCVal(
+            type=messages.StellarSCValType.SCV_I32, i32=val.i32.int32
+        )
+    elif val.type == xdr.SCValType.SCV_U64:
+        assert val.u64 is not None
+        return messages.StellarSCVal(
+            type=messages.StellarSCValType.SCV_U64, u64=val.u64.uint64
+        )
+    elif val.type == xdr.SCValType.SCV_I64:
+        assert val.i64 is not None
+        return messages.StellarSCVal(
+            type=messages.StellarSCValType.SCV_I64, i64=val.i64.int64
+        )
+    elif val.type == xdr.SCValType.SCV_TIMEPOINT:
+        assert val.timepoint is not None
+        return messages.StellarSCVal(
+            type=messages.StellarSCValType.SCV_TIMEPOINT,
+            timepoint=val.timepoint.time_point.uint64,
+        )
+    elif val.type == xdr.SCValType.SCV_DURATION:
+        assert val.duration is not None
+        return messages.StellarSCVal(
+            type=messages.StellarSCValType.SCV_DURATION,
+            duration=val.duration.duration.uint64,
+        )
+    elif val.type == xdr.SCValType.SCV_U128:
+        assert val.u128 is not None
+        return messages.StellarSCVal(
+            type=messages.StellarSCValType.SCV_U128,
+            u128=messages.StellarUInt128Parts(
+                hi=val.u128.hi.uint64, lo=val.u128.lo.uint64
+            ),
+        )
+    elif val.type == xdr.SCValType.SCV_I128:
+        assert val.i128 is not None
+        return messages.StellarSCVal(
+            type=messages.StellarSCValType.SCV_I128,
+            i128=messages.StellarInt128Parts(
+                hi=val.i128.hi.int64, lo=val.i128.lo.uint64
+            ),
+        )
+    elif val.type == xdr.SCValType.SCV_U256:
+        assert val.u256 is not None
+        return messages.StellarSCVal(
+            type=messages.StellarSCValType.SCV_U256,
+            u256=messages.StellarUInt256Parts(
+                hi_hi=val.u256.hi_hi.uint64,
+                hi_lo=val.u256.hi_lo.uint64,
+                lo_hi=val.u256.lo_hi.uint64,
+                lo_lo=val.u256.lo_lo.uint64,
+            ),
+        )
+    elif val.type == xdr.SCValType.SCV_I256:
+        assert val.i256 is not None
+        return messages.StellarSCVal(
+            type=messages.StellarSCValType.SCV_I256,
+            i256=messages.StellarInt256Parts(
+                hi_hi=val.i256.hi_hi.int64,
+                hi_lo=val.i256.hi_lo.uint64,
+                lo_hi=val.i256.lo_hi.uint64,
+                lo_lo=val.i256.lo_lo.uint64,
+            ),
+        )
+    elif val.type == xdr.SCValType.SCV_BYTES:
+        assert val.bytes is not None
+        return messages.StellarSCVal(
+            type=messages.StellarSCValType.SCV_BYTES, bytes=val.bytes.sc_bytes
+        )
+    elif val.type == xdr.SCValType.SCV_STRING:
+        assert val.str is not None
+        return messages.StellarSCVal(
+            type=messages.StellarSCValType.SCV_STRING,
+            string=val.str.sc_string,  # raw bytes, not necessarily UTF-8
+        )
+    elif val.type == xdr.SCValType.SCV_SYMBOL:
+        assert val.sym is not None
+        return messages.StellarSCVal(
+            type=messages.StellarSCValType.SCV_SYMBOL,
+            symbol=val.sym.sc_symbol.decode("utf-8"),
+        )
+    elif val.type == xdr.SCValType.SCV_VEC:
+        # SCV_VEC's vector is an XDR pointer (SCVec*), i.e. technically nullable,
+        # but a null vector is not a valid Soroban value. Reject it so we never
+        # coerce Vec(None) into Vec([]) and end up signing different bytes than
+        # the input XDR (the firmware always encodes the vector as present).
+        if val.vec is None:
+            raise ValueError("SCV_VEC with a null vector is not supported")
+        return messages.StellarSCVal(
+            type=messages.StellarSCValType.SCV_VEC,
+            vec=[_read_sc_val(v) for v in val.vec.sc_vec],
+        )
+    elif val.type == xdr.SCValType.SCV_MAP:
+        # SCV_MAP's map is an XDR pointer (SCMap*); same reasoning as SCV_VEC.
+        if val.map is None:
+            raise ValueError("SCV_MAP with a null map is not supported")
+        return messages.StellarSCVal(
+            type=messages.StellarSCValType.SCV_MAP,
+            map=[
+                messages.StellarSCValMapEntry(
+                    key=_read_sc_val(item.key), value=_read_sc_val(item.val)
+                )
+                for item in val.map.sc_map
+            ],
+        )
+    elif val.type == xdr.SCValType.SCV_ADDRESS:
+        assert val.address is not None
+        return messages.StellarSCVal(
+            type=messages.StellarSCValType.SCV_ADDRESS,
+            address=_read_sc_address(val.address),
+        )
+    else:
+        raise ValueError(f"Unsupported SCVal type: {val.type}")
+
+
+def _read_invoke_contract_args(
+    data: xdr.InvokeContractArgs,
+) -> messages.StellarInvokeContractArgs:
+    """Read InvokeContractArgs from XDR."""
+    return messages.StellarInvokeContractArgs(
+        contract_address=_read_sc_address(data.contract_address),
+        function_name=data.function_name.sc_symbol.decode("utf-8"),
+        args=[_read_sc_val(arg) for arg in data.args],
+    )
+
+
+def _read_authorized_function(
+    function: xdr.SorobanAuthorizedFunction,
+) -> messages.StellarSorobanAuthorizedFunction:
+    """Read SorobanAuthorizedFunction from XDR."""
+    if (
+        function.type
+        == xdr.SorobanAuthorizedFunctionType.SOROBAN_AUTHORIZED_FUNCTION_TYPE_CONTRACT_FN
+    ):
+        assert function.contract_fn is not None
+        return messages.StellarSorobanAuthorizedFunction(
+            type=messages.StellarSorobanAuthorizedFunctionType.SOROBAN_AUTHORIZED_FUNCTION_TYPE_CONTRACT_FN,
+            contract_fn=_read_invoke_contract_args(function.contract_fn),
+        )
+    else:
+        raise ValueError(f"Unsupported SorobanAuthorizedFunction type: {function.type}")
+
+
+def _read_address_credentials(
+    address_credentials: xdr.SorobanAddressCredentials,
+) -> messages.StellarSorobanAddressCredentials:
+    """Read SorobanAddressCredentials from XDR."""
+    return messages.StellarSorobanAddressCredentials(
+        address=_read_sc_address(address_credentials.address),
+        nonce=address_credentials.nonce.int64,
+        signature_expiration_ledger=address_credentials.signature_expiration_ledger.uint32,
+        signature=_read_sc_val(address_credentials.signature),
+    )
+
+
+def _read_delegate_signature(
+    delegate: xdr.SorobanDelegateSignature,
+) -> messages.StellarSorobanDelegateSignature:
+    """Read SorobanDelegateSignature from XDR."""
+    return messages.StellarSorobanDelegateSignature(
+        address=_read_sc_address(delegate.address),
+        signature=_read_sc_val(delegate.signature),
+        nested_delegates=[
+            _read_delegate_signature(nested) for nested in delegate.nested_delegates
+        ],
+    )
+
+
+def _read_address_credentials_with_delegates(
+    credentials: xdr.SorobanAddressCredentialsWithDelegates,
+) -> messages.StellarSorobanAddressCredentialsWithDelegates:
+    """Read SorobanAddressCredentialsWithDelegates from XDR."""
+    return messages.StellarSorobanAddressCredentialsWithDelegates(
+        address_credentials=_read_address_credentials(credentials.address_credentials),
+        delegates=[
+            _read_delegate_signature(delegate) for delegate in credentials.delegates
+        ],
+    )
+
+
+def _read_credentials(
+    credentials: xdr.SorobanCredentials,
+) -> messages.StellarSorobanCredentials:
+    """Read SorobanCredentials from XDR."""
+    if (
+        credentials.type
+        == xdr.SorobanCredentialsType.SOROBAN_CREDENTIALS_SOURCE_ACCOUNT
+    ):
+        return messages.StellarSorobanCredentials(
+            type=messages.StellarSorobanCredentialsType.SOROBAN_CREDENTIALS_SOURCE_ACCOUNT
+        )
+    elif (
+        HAVE_STELLAR_SDK_PROTOCOL_27
+        and credentials.type
+        == xdr.SorobanCredentialsType.SOROBAN_CREDENTIALS_ADDRESS_V2
+    ):
+        assert credentials.address_v2 is not None
+        return messages.StellarSorobanCredentials(
+            type=messages.StellarSorobanCredentialsType.SOROBAN_CREDENTIALS_ADDRESS_V2,
+            address_v2=_read_address_credentials(credentials.address_v2),
+        )
+    elif (
+        HAVE_STELLAR_SDK_PROTOCOL_27
+        and credentials.type
+        == xdr.SorobanCredentialsType.SOROBAN_CREDENTIALS_ADDRESS_WITH_DELEGATES
+    ):
+        assert credentials.address_with_delegates is not None
+        return messages.StellarSorobanCredentials(
+            type=messages.StellarSorobanCredentialsType.SOROBAN_CREDENTIALS_ADDRESS_WITH_DELEGATES,
+            address_with_delegates=_read_address_credentials_with_delegates(
+                credentials.address_with_delegates
+            ),
+        )
+    else:
+        # The legacy, to-be-deprecated SOROBAN_CREDENTIALS_ADDRESS is
+        # intentionally not supported.
+        raise ValueError(f"Unsupported SorobanCredentials type: {credentials.type}")
+
+
+def _read_authorized_invocation(
+    invocation: xdr.SorobanAuthorizedInvocation,
+) -> messages.StellarSorobanAuthorizedInvocation:
+    """Read SorobanAuthorizedInvocation from XDR."""
+    return messages.StellarSorobanAuthorizedInvocation(
+        function=_read_authorized_function(invocation.function),
+        sub_invocations=[
+            _read_authorized_invocation(sub) for sub in invocation.sub_invocations
+        ],
+    )
+
+
+def _read_authorization_entry(
+    entry: xdr.SorobanAuthorizationEntry,
+) -> messages.StellarSorobanAuthorizationEntry:
+    """Read SorobanAuthorizationEntry from XDR."""
+    return messages.StellarSorobanAuthorizationEntry(
+        credentials=_read_credentials(entry.credentials),
+        root_invocation=_read_authorized_invocation(entry.root_invocation),
+    )
+
+
+def _read_host_function(
+    host_function: xdr.HostFunction,
+) -> messages.StellarHostFunction:
+    """Read HostFunction from XDR."""
+    if host_function.type != xdr.HostFunctionType.HOST_FUNCTION_TYPE_INVOKE_CONTRACT:
+        raise ValueError(f"Unsupported host function type: {host_function.type}")
+
+    assert host_function.invoke_contract is not None
+    return messages.StellarHostFunction(
+        type=messages.StellarHostFunctionType.HOST_FUNCTION_TYPE_INVOKE_CONTRACT,
+        invoke_contract=_read_invoke_contract_args(host_function.invoke_contract),
+    )
